applications/speech_to_text_llm/requirements.txt
openai==0.27.2
openai-whisper==20230314